+++ ft_strdup.c
@@ -6,7 +6,7 @@
 /*   By: cvernius <cvernius@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/12 20:58:22 by cvernius          #+#    #+#             */
-/*   Updated: 2019/04/24 10:25:05 by cvernius         ###   ########.fr       */
+/*   Updated: 2019/05/03 14:18:24 by cvernius         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,7 +15,7 @@
 char	*ft_strdup(const char *s1)
 {
 	char	*str2;
-	int		i;
+	size_t	i;
 
 	i = 0;
 	if ((str2 = (char *)malloc(sizeof(char) * (ft_strlen(s1) + 1))))
