+++ ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: cvernius <cvernius@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/04/23 20:54:25 by cvernius          #+#    #+#             */
-/*   Updated: 2019/04/30 21:07:09 by cvernius         ###   ########.fr       */
+/*   Updated: 2019/05/03 14:09:06 by cvernius         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,8 +14,8 @@
 
 char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 {
-	unsigned int	i;
-	char			*str;
+	size_t		i;
+	char		*str;
 
 	i = 0;
 	str = (char*)s;
